% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jakiPrzebieg.R
\name{jakiPrzebieg}
\alias{jakiPrzebieg}
\title{Sredni przebieg aut wyprodukowanych w danym roku}
\usage{
jakiPrzebieg(rok = "2012", auta)
}
\arguments{
\item{rok}{Rok, dla ktorego liczona ma byc srednia}

\item{auta}{Zbior danych do liczenia sredniej.}
}
\description{
Funkcja jakiPrzebieg() wyznacza sredni przebieg aut,
ktore jako data produkcji maja podany wskazany rok.
}
