% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jakiPrzebieg.R
\name{jakiPrzebieg}
\alias{jakiPrzebieg}
\title{Średni przebieg aut wyprodukowanych w danym roku}
\usage{
jakiPrzebieg(rok = "2012", auta)
}
\arguments{
\item{rok}{Rok, dla którego liczona ma być średnia, domyślnie 2012.}

\item{auta}{Zbiór danych, na bazie którego liczona ma być średnia.}
}
\description{
Funkcja jakiPrzebieg() wyznacza średni przebieg aut,
które jako data produkcji mają podany wskazany rok.
}

